Fix comments and package naming
