+++ src/main/java/seedu/address/logic/parser/DeleteDateCommandParser.java
@@ -12,7 +12,7 @@ public class DeleteDateCommandParser implements Parser<DeleteDateCommand> {
     /**
      * Parses the given {@code String} of arguments in the context of the DeleteDateCommand
      * and returns a DeleteDateCommand object for execution.
-     * @throws ParseException if the user input does not conform the expected format
+     * @throws ParseException if the user input does not conform to the expected format
      */
     public DeleteDateCommand parse(String args) throws ParseException {
         try {
@@ -20,7 +20,6 @@ public class DeleteDateCommandParser implements Parser<DeleteDateCommand> {
             return new DeleteDateCommand(index);
         } catch (ParseException pe) {
             throw new ParseException(
-                // String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteDateCommand.MESSAGE_USAGE), pe);
                 String.format(pe.getMessage(), DeleteDateCommand.MESSAGE_USAGE), pe);
         }
     }
