+++ src/main/java/seedu/address/model/date/exceptions/DuplicateImportantDateException.java
@@ -2,7 +2,7 @@ package seedu.address.model.date.exceptions;
 
 /**
  * Signals that the operation will result in duplicate ImportantDates (ImportantDates are considered duplicates if they
- * have the same description and details).
+ * have the same description).
  */
 public class DuplicateImportantDateException extends RuntimeException {
     public DuplicateImportantDateException() {
