+++ src/main/java/seedu/address/model/lesson/Time.java
@@ -7,6 +7,7 @@ public class Time implements Comparable<Time> {
 
     public static final String VALIDATION_REGEX = "^(0[0-9]|1[0-9]|2[0-3])[0-5][0-9]$";
     public static final String MESSAGE_CONSTRAINTS = "Time should be in HHMM format, from 0000 to 2359";
+
     public final String timeOfTuition;
 
     /**
