+++ src/main/java/seedu/address/storage/StorageManager.java
@@ -14,7 +14,7 @@ import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.UserPrefs;
 
 /**
- * Manages storage of AddressBook data in local storage.
+ * Manages storage of TutorsPet data in local storage.
  */
 public class StorageManager implements Storage {
