+++ src/main/java/seedu/address/ui/ImportantDatesWindow.java
@@ -7,7 +7,7 @@ import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.logic.Logic;
-import seedu.address.ui.PanelList.DateListPanel;
+import seedu.address.ui.panel.DateListPanel;
 
 /**
  * Controller for the Important Dates page
