+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -21,7 +21,7 @@ import seedu.address.logic.Logic;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.ui.PanelList.PersonListPanel;
+import seedu.address.ui.panel.PersonListPanel;
 
 /**
  * The Main Window. Provides the basic application layout containing
