+++ src/main/java/seedu/address/ui/ScheduleWindow.java
@@ -9,7 +9,7 @@ import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.logic.Logic;
-import seedu.address.ui.PanelList.LessonListPanel;
+import seedu.address.ui.panel.LessonListPanel;
 
 /**
  * Controller for a schedule page
