+++ src/main/java/seedu/address/ui/card/DateCard.java
@@ -1,19 +1,19 @@
-package seedu.address.ui.Card;
+package seedu.address.ui.card;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
-import seedu.address.model.lesson.Lesson;
+import seedu.address.model.date.ImportantDate;
 import seedu.address.ui.UiPart;
 
 
 /**
  * An UI component that displays information of an {@code Important Date}.
  */
-public class LessonListCard extends UiPart<Region> {
+public class DateCard extends UiPart<Region> {
 
-    private static final String FXML = "LessonListCard.fxml";
+    private static final String FXML = "DateListCard.fxml";
 
     /**
      * Note: Certain keywords such as "location" and "resources" are reserved keywords in JavaFX.
@@ -23,23 +23,26 @@ public class LessonListCard extends UiPart<Region> {
      * @see <a href="https://github.com/se-edu/addressbook-level4/issues/336">The issue on AddressBook level 4</a>
      */
 
-    public final Lesson lesson;
+    public final ImportantDate importantDate;
 
     @javafx.fxml.FXML
     private HBox cardPane;
     @FXML
-    private Label lessonDetails;
+    private Label description;
     @FXML
-    private Label personNames;
+    private Label details;
+    @FXML
+    private Label id;
 
     /**
-     * Creates a {@code LessonListCard} with the given {@code Lesson} and index to display.
+     * Creates a {@code DateCard} with the given {@code Important Date} and index to display.
      */
-    public LessonListCard(Lesson lesson) {
+    public DateCard(ImportantDate importantDate, int displayedIndex) {
         super(FXML);
-        this.lesson = lesson;
-        lessonDetails.setText(lesson.getTimeInString());
-        personNames.setText(lesson.getPersonInString());
+        this.importantDate = importantDate;
+        id.setText(displayedIndex + ". ");
+        description.setText(importantDate.getDescription().description);
+        details.setText(importantDate.getDetails().toString());
     }
 
     @Override
@@ -50,14 +53,14 @@ public class LessonListCard extends UiPart<Region> {
         }
 
         // instanceof handles nulls
-        if (!(other instanceof LessonListCard)) {
+        if (!(other instanceof DateCard)) {
             return false;
         }
 
         // state check
-        LessonListCard card = (LessonListCard) other;
-        return lessonDetails.getText().equals(card.lessonDetails.getText())
-                && personNames.getText().equals(card.personNames.getText());
+        DateCard card = (DateCard) other;
+        return id.getText().equals(card.id.getText())
+                && importantDate.equals(card.importantDate);
     }
 
 }
